Make text slightly darker and quotes non-italic
The text "looks" nicer when it's a little lighter, but it breaks
readability. I've made it closer to full black now so it can be more
readable.

The quotes being italic I felt was a little weird. The lighter color,
left border and padding would probably be enough (not to mention the
semantics <blockquote> itself).