CSS: Smaller headers - adjusted scale progression
Adding YET ANOTHER variable for something should just be left to the web
user-agent stylesheet. I'm so extra, I know.

Anyways in this commit I've adjusted the scaling for h1, h2, and h3 so
they aren't too big.

Reasoning:
  Big = attention-seeking             = ugly = bad = bad design
  Big = wraps text on smaller screens = ugly = bad = bad design

The scale progression of headers is inherited from simple.css. In the
future I'll probably remove the font-size settings altogether and leave
it up to the user to set.